+++ src/components/Chatting/MessagesBody/ChatDate.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import theme from "styles/theme";
+import theme from "tools/theme";
 import DottedLine from "components/common/DottedLine";
 
 const ChatDate = (props) => {
