+++ src/components/Chatting/MessagesBody/ChatPaySettle.tsx
@@ -1,4 +1,4 @@
-import theme from "styles/theme";
+import theme from "tools/theme";
 import CreditCardRoundedIcon from "@mui/icons-material/CreditCardRounded";
 import SendRoundedIcon from "@mui/icons-material/SendRounded";
