+++ src/components/Chatting/SideChat.jsx
@@ -1,6 +1,6 @@
 import PropTypes from "prop-types";
 import Chatting from "components/Chatting/Chatting";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 const SideChat = ({ roomId }) => {
   const styleSideChat = {
