+++ src/components/Mypage/PopupReport/ReportOption.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from "react-i18next";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 export type ReportOptionType = "Reporting" | "Reported";
