+++ src/components/Myroom/R1Myroom.tsx
@@ -5,7 +5,7 @@ import RLayout from "components/common/RLayout";
 import Pagination, {
   PAGE_MAX_ITEMS,
 } from "components/common/pagination/Pagination";
-import theme from "styles/theme";
+import theme from "tools/theme";
 import Empty from "components/common/Empty";
 
 /**
