+++ src/components/Reporting/PopupReport.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, FormEvent, useState } from "react";
 import ProfileImg from "components/Mypage/ProfileImg";
 import Modal from "components/common/modal/Modal";
 import { useAxios } from "hooks/useTaxiAPI";
-import theme from "styles/theme";
+import theme from "tools/theme";
 import { useSetRecoilState } from "recoil";
 import alertAtom from "atoms/alert";
 import regExpTest from "tools/regExpTest";
