+++ src/components/Skeleton/AlertProvider.tsx
@@ -3,7 +3,7 @@ import Modal from "components/common/modal/Modal";
 import { useRecoilState } from "recoil";
 import alertAtom from "atoms/alert";
 import Button from "components/common/Button";
-import theme from "styles/theme";
+import theme from "tools/theme";
 import DottedLine from "components/common/DottedLine";
 
 import WbIncandescentRoundedIcon from "@material-ui/icons/WbIncandescentRounded";
