+++ src/components/Skeleton/Navigation.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { Link } from "react-router-dom";
 import { useLocation } from "react-router-dom";
 import hoverEventSet from "tools/hoverEventSet";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 import { useRecoilValue } from "recoil";
 import isVirtualKeyboardDetectedAtom from "atoms/isVirtualKeyboardDetected";
