+++ src/components/common/DottedLine.tsx
@@ -1,4 +1,4 @@
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 type Direction = "row" | "column";
