+++ src/components/common/HeaderBar.tsx
@@ -1,4 +1,4 @@
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 const HeaderBar = () => {
   return (
