+++ src/components/common/Title.jsx
@@ -1,7 +1,7 @@
 import { memo } from "react";
 import PropTypes from "prop-types";
 import RLayout from "components/common/RLayout";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 import LocalTaxiRoundedIcon from "@mui/icons-material/LocalTaxiRounded";
 import SearchRoundedIcon from "@mui/icons-material/SearchRounded";
