+++ src/components/common/Tooltip.tsx
@@ -1,4 +1,4 @@
-import theme from "styles/theme";
+import theme from "tools/theme";
 import Tooltip from "@mui/material/Tooltip";
 
 type ToolTipProps = {
