+++ src/components/common/pagination/Pagination.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from "react";
 import { useLocation, useHistory } from "react-router-dom";
 import qs from "qs";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 import KeyboardArrowLeftIcon from "@mui/icons-material/KeyboardArrowLeftRounded";
 import KeyboardArrowRightIcon from "@mui/icons-material/KeyboardArrowRightRounded";
