+++ src/components/common/room/Room.jsx
@@ -5,7 +5,7 @@ import { getLocationName } from "tools/trans";
 import { date2str } from "tools/moment";
 import loginInfoDetailAtom from "atoms/loginInfoDetail";
 import DottedLine from "components/common/DottedLine";
-import theme from "styles/theme";
+import theme from "tools/theme";
 import "./Room.css";
 
 import ArrowRightAltRoundedIcon from "@mui/icons-material/ArrowRightAltRounded";
