+++ src/components/common/roomOptions/MaxPeople.tsx
@@ -1,6 +1,6 @@
 import { memo } from "react";
 import WhiteContainer from "components/common/WhiteContainer";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 import PeopleRoundedIcon from "@material-ui/icons/PeopleRounded";
 import RemoveRoundedIcon from "@material-ui/icons/RemoveRounded";
