+++ src/components/common/roomOptions/Name.tsx
@@ -1,6 +1,6 @@
 import { memo } from "react";
 import WhiteContainer from "components/common/WhiteContainer";
-import theme from "styles/theme";
+import theme from "tools/theme";
 
 import EditRoundedIcon from "@material-ui/icons/EditRounded";
